Part B Completion (pending more testing)
Completed the conditional cases for part b for constant scanning.

Initial issue on line 98 was resolved. Cause of problem was "\n" escape was included in dictionary values (FILEHASH\n), resolved issue using strip() function.

Checking if reference file was deleted needs to be done at the start of the def due to the reference file being used to update key/value pairs. If file does not exist, user is notified and a return statement sends user back to main menu.

To check if file has been created I used the "contains" function from the dictionary .
LOGIC (Line 103):
if dictionary does not contain the filepath (Key) of the patients in patients folder, it means that a new file has been added to the patients folder.

Fixing the initial problem in Line 98 resulted in the initial "file has been changed" Test to work as expected

I have kept Line 117-122 active as It does nothing yet. it is advised to be removed.

Added Return statements to file integrity checks so every outcome will return to the main menu