Began Work on Part C
Began Work on Part C

Part C has a an additional selection sub menu that lets the user see if the specific file they have selected is a different version

The different version will be identified  by saving the each files last save date into a separate file (much like reference file).  The date-reference file will compare individual files by their "last modified date", if the save date  is different it will mean there is a new version.  (I have created a save date reference file and moved it into the reference file folder)

What still needs to be done is the creation of a key/value pair system using the filepath and the last save date (Similar to part B).

What can be implemented in the future is to utilize the capabilities of python to detect when a file is saved by continuously monitoring the "os.path.getmtime()" function. This dictates the modified last date whenever a file is saved. It might be possible to save the previous version of a file in another directory, but this still requires some research.