Part C completed
Completed Part C:

Utilized the same concept as part B using a dictionary with the datereference file in the referencefile folder.

We compare if the dates are the same and if the dates are different it will send an alert notifying the user that the version in the patients folder is a different version.

Demo of all parts will be conducted on Sunday Sept 9