Option 2 - Client side implementation
Created new Alert html page. Created boolean checkers that display message in alert html when a condition is met in option 2.

Error caused when checkFile() is called as checkFile has not been defined as of yet this early in the script.

Will have to implement an asynchronous method call to fix the issue (To do Sunday, 24)

Co-Authored-By: Amanda Jose <16587142+amanda-jose@users.noreply.github.com>